chore: duplicate test in android and jvm
